melhorias no app

+++ app/src/main/java/com/example/rako/bankingapp/model/Ingredient.java
@@ -19,7 +19,7 @@ public class Ingredient implements Parcelable{
     }
 
     protected Ingredient(Parcel in) {
-        quantity = in.readInt();
+        quantity = in.readFloat();
         measure = in.readString();
         ingredient = in.readString();
     }
